Fetch Oracle from MTG JSON
Instead of bundling the json object, fetch it from the source.

We follow these rules:
* On install or update from 3.6.X, create a new IndexedDB and set
storage.local defaults
* On startup, check to see if it's been over two weeks since a check. If
yes, download the MTG JSON version file and compare it to the last
version we download. If different, bump storage.local.version
* When storage.local.version changes, download the Oracle json data

This ensures we download the full file only when necessary.